+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Issue #11] onComplete should accept functions that return Promise
 - [Issue #10] BootstrapFunction should accept return type void or Promise<void>
 - [Issue #9] Add bootstrap class for cleaner extension
+- [Issue #8] Add teardown method and handlers
 - [Issue #7] Add onFinally method
 
 ### Fixes